requirements.txt
# Core NLP
nltk==3.8.1
keybert==0.8.3
scikit-learn==1.4.0

# PDF Reading (primary: pdfplumber, fallback: pypdf)
pdfplumber==0.11.0
pypdf==4.2.0

# GenAI
google-genai==1.5.0

# Utilities
python-dotenv==1.0.1
colorama==0.4.6

# Optional: export to PDF
fpdf2==2.7.9